% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{ar1}
\alias{ar1}
\title{Compute first autocorrelation (AR1) of a time series}
\usage{
ar1(ts)
}
\arguments{
\item{ts}{time series}
}
\value{
value of autocorrelation of lag 1
}
\description{
Compute first autocorrelation (AR1) of a time series
}
\examples{
data(baltic)
ar1(baltic$Cod)
}
\seealso{
\code{\link{infoTS}}
}
