% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baltic}
\alias{baltic}
\title{Dynamics of Baltic Sea ecosystem}
\format{A data frame with 32 rows (corresponding to 32 years) and 14 variables:
\describe{
  \item{Cod}{Cod (Gadus morhua) spawner biomass, in tonnes}
  \item{Herring}{Herring (Clupea harengus) spawner biomass, in tonnes}
  \item{Sprat}{Sprat (Sprattus sprattus) spawner biomass, in tonnes}
  \item{Acartia}{Acartia spp. biomass in Gotland basin in spring, in mg.m-3}
  \item{Pseudo}{Pseudocalanus acuspes biomass in Gotland basin in spring, in mg.m-3}
  \item{Temora}{Temora longicornis biomass in Gotland basin in spring, in mg.m-3}
  \item{Chla}{Chlorophyll a concentration in Gotland basin in spring, in mg.m-3}
  \item{Diatom}{Diatoms in Gotland basin in spring, in mg.m-3}
  \item{Dino}{Dinoflagellates concentration in Gotland basin in spring, in mg.m-3}
  \item{Cyano}{Cyanobacteria in Gotland basin in spring, in mg.m-3}
  \item{SST}{Sea surface temperature in Gotland basin in spring, in degree Celsius}
  \item{SSS}{Sea surface salinity in Gotland basin in spring, in psu}
  \item{O2}{Deep oxygen concentration in Gotland basin in spring, in mL.L-1}
  \item{Anoxic}{Anoxic area}
}}
\source{
ICES/HELCOM Working Group on Integrated Assessments of the Baltic Sea \url{https://www.ices.dk/community/groups/Pages/WGIAB.aspx}
}
\usage{
baltic
}
\description{
A dataset containing time series of multiple components of 
the Gotland basin ecosystem in spring between 1980 and 2011.
}
\references{
Möllmann C., Diekmann R. , Müller-Karulis B. , Kornilovs G. , Plikshs M. and Axe P. (2009), Reorganization of a large marine ecosystem due to atmospheric and anthropogenic pressure: a discontinuous regime shift in the Central Baltic Sea. Global Change Biology, 15: 1377-1393. doi:10.1111/j.1365-2486.2008.01814.x
}
\keyword{datasets}
