% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comita.R
\docType{package}
\name{comita}
\alias{comita}
\title{Comparative tools for Ecosystem Integrative Trend Analysis}
\description{
Integrated trend analyses (ITA) are commonly applied to summarise changes that have 
occurred in recent decades in ecosystems. Currently there are multiple methods used 
for ITA, raising question on which is the best method to use for ITA. The aim of this 
application is to provide a tool to easily compare different ITA methods, and test 
the effect of data pre-processing and variables selection on the final results. 
Hopefully, this package will raise awareness of ecologists to different dimension 
reduction techniques, as well as provide reason and confidence on the choice of ITA 
methods according to the characteristics of their dataset.
}
\details{
comita: Comparative tools for Ecosystem Integrative Trend Analysis
}
