% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{cor_rand}
\alias{cor_rand}
\title{Compute p-values for correlation between time series}
\usage{
cor_rand(ts1, ts2, nrep = 100, metrand = "phase", ...)
}
\arguments{
\item{ts1}{single time series}

\item{ts2}{single time series}

\item{nrep}{number of repetition}

\item{metrand}{define the method to create random dataset (phase for keeping phase structure, random data, or random + trend (default = TRUE)}

\item{...}{elements sent to function cor()}
}
\value{
p-value of correlation coefficient
}
\description{
Compute p-values for correlation between time series
}
\details{
Variable \code{metrand} can be "phase", "rand" or "trend": 
\itemize{
\item \code{phase} keep similar phase structure in the dataset
\item \code{rand} randomly generated data with similar mean and sd
\item \code{trend} randomly generated data with similar trend
}
}
\examples{
data(baltic)
# test the correlation between cod and herring biomass
# using surrogates data conserving phase structure
cor_rand(baltic$Cod, baltic$Herring)

#Compare to a traditional test
cor.test(baltic$Cod, baltic$Herring)
}
\seealso{
\code{\link{createRand}}
}
