% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{createRand}
\alias{createRand}
\title{Create a random dataset with similar characteristics than original data}
\usage{
createRand(dat, ar = TRUE, tr = FALSE)
}
\arguments{
\item{dat}{matrix of time series (time in rows, variables in columns)}

\item{ar}{whether to keep the autocorrelation structure (default = TRUE)}

\item{tr}{whether to add a trend similar to the original one (default =FALSE)}
}
\value{
a randomly generated matrix of similar size than \code{dat}
}
\description{
Create a random dataset with similar characteristics than original data
}
\seealso{
\code{\link{ita.rand}}, \code{\link{ita.nrand}}
}
