% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{getpvar}
\alias{getpvar}
\title{Function to know the percentage of variance explained}
\usage{
getpvar(ts, co, dat, npc)
}
\arguments{
\item{ts}{scores of the time series on the new dimensions}

\item{co}{scores of the variables on the new dimensions}

\item{dat}{matrix with}

\item{npc}{number of new dimensions}
}
\value{
the cumulated percentage of variance explained by the successive new dimensions
}
\description{
Function to know the percentage of variance explained
}
\note{
inspired from https://ro-che.info/articles/2017-12-11-pca-explained-variance
}
