% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{infoTS}
\alias{infoTS}
\title{Return a set of statistics on the original dataset}
\usage{
infoTS(dat)
}
\arguments{
\item{dat}{matrix of time series (time in rows, variables in columns)}
}
\value{
characteristics of the dataset, including
\describe{
  \item{nbio}{number of biological time series}
  \item{ntime}{length of the time series}
  \item{meanAR}{mean AR1 of the time series}
  \item{maxAR}{maximum AR1 among the different time series}
  \item{meanCor}{mean pairwize correlation among time series}
  \item{maxCor}{maximum pairwize correlation among time series}
}
}
\description{
Return a set of statistics on the original dataset
}
\examples{
data(baltic)
infoTS(baltic)
}
\seealso{
\code{\link{ar1}}
}
