% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita.R
\name{ita.dpca}
\alias{ita.dpca}
\title{Run a Dynamic Principal Component Analysis}
\usage{
ita.dpca(dat, npc)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of processes}
}
\value{
Output of ITA with 
\itemize{
\item \code{ts} scores of the time on the principal component
\item \code{co} scores of the variables on the principal component
\item \code{eig} eigen values
\item \code{dat} original dataset
\item \code{npc} number of selected principal component
\item \code{mita} method used for integrated trend analysis 
}
}
\description{
Run a Dynamic Principal Component Analysis
}
\note{
lag set to 1, which correspond to AR1 processes
Function based on dpca from freqdom package
}
\references{
Ku W, Storer RH, Georgakis C (1995) Disturbance detection and isolation by dynamic principal component analysis. Chemom Intell Lab Syst 30:179–196
Ketelaere B De, Hubert M, Schmitt E (2015) Overview of PCA-based statistical process-monitoring methods for time-dependent, high-dimensional data. J Qual Technol 47:318–335
}
