% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita.R
\name{ita.fca}
\alias{ita.fca}
\title{Run Forecastable Component Analysis}
\usage{
ita.fca(dat, npc)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of new dimensions}
}
\value{
Output of ITA with 
\describe{
\item \code{ts} scores of the time series on the new dimensions 
\item \code{co} NULL - no scores of the variables on the new dimensions 
\item \code{eig} NULL 
\item \code{dat} original dataset
\item \code{npc} number of new dimensions
\item \code{mita} method used for integrated trend analysis 
}
}
\description{
Run Forecastable Component Analysis
}
\note{
Use function foreca from the ForeCA package.
Limited to matrices with more time steps than variables
}
\examples{

}
\references{
Goerg GM (2013) Forecastable Component Analysis. In: Proceedings of the 30th International Conference on Machine Learning.p 64–72
}
