% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita.R
\name{ita.ica}
\alias{ita.ica}
\title{Run Independent Component Analysis}
\usage{
ita.ica(dat, npc)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of new dimensions}
}
\value{
Output of ITA with 
\describe{
\item \code{ts} scores of the time series on the new dimensions 
\item \code{co} NULL - no scores of the variables on the new dimensions 
\item \code{eig} NULL 
\item \code{dat} original dataset
\item \code{npc} number of new dimensions
\item \code{mita} method used for integrated trend analysis 
}
}
\description{
Run Independent Component Analysis
}
\note{
Use function fastICA from the fastICA package.
}
\examples{

}
\references{
Back AD, Weigend AS (1997) A First Application of Independent Component Analysis to Extracting Structure from Stock Returns. Int J Neural Syst 08:473–484
Blaschke T, Berkes P, Wiskott L (2006) What Is the Relation Between Slow Feature Analysis and Independent Component Analysis? Neural Comput 18:2495–2508
}
