% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita.R
\name{ita.lle}
\alias{ita.lle}
\title{Run Local Linear Embedding}
\usage{
ita.lle(dat, npc)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of new dimensions}
}
\value{
Output of ITA with 
\itemize{
\item \code{ts} scores of the time series on the new dimensions 
\item \code{co} NULL - no scores of the variables on the new dimensions 
\item \code{eig} NULL 
\item \code{dat} original dataset
\item \code{npc} number of new dimensions
\item \code{mita} method used for integrated trend analysis 
}
}
\description{
Run Local Linear Embedding
}
\note{
Use function lle from the lle package.
}
\references{
Roweis ST (2000) Nonlinear Dimensionality Reduction by Locally Linear Embedding. Science (80- ) 290:2323–2326
}
