% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita.R
\name{ita.mafa}
\alias{ita.mafa}
\title{Run a Min/Max Autocorrelation Factors Analysis (MAFA)}
\usage{
ita.mafa(dat, npc)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of processes}
}
\value{
Output of ITA with 
\itemize{
\item \code{ts} scores of the time on the principal component
\item \code{co} scores of the variables on the principal component
\item \code{eig} eigen values
\item \code{dat} original dataset
\item \code{npc} number of selected principal component
\item \code{mita} method used for integrated trend analysis 
}
}
\description{
Run a Min/Max Autocorrelation Factors Analysis (MAFA)
}
\note{
Use maf function created by M.Woillez and J. Rivoirard.
}
\references{
Solow AR (1994) Detecting Change in the Composition of a Multispecies Community. Biometrics 50:556
Woillez M, Rivoirard J, Petitgas P (2009) Using min/max autocorrelation factors of survey-based indicators to follow the evolution of fish stocks in time. Aquat Living Resour 22:193–200
}
\seealso{
\code{\link{maf}}
}
