% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita.R
\name{ita.mds}
\alias{ita.mds}
\title{Run Multidimensional scaling}
\usage{
ita.mds(dat, npc)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of processes}
}
\value{
Output of ITA with 
\describe{
\item \code{ts} scores of the time series on the new dimensions 
\item \code{co} NULL - no scores of the variables on the new dimensions 
\item \code{eig} NULL 
\item \code{dat} original dataset
\item \code{npc} number of selected principal component
\item \code{mita} method used for integrated trend analysis 
}
}
\description{
Run Multidimensional scaling
}
\examples{

}
