% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{ita.nrand}
\alias{ita.nrand}
\title{Extension of ita.rand to generate multiple datasets}
\usage{
ita.nrand(dat, npc = 2, met = 1, sca = TRUE, logt = FALSE,
  metrand = "phase", nrep = 10)
}
\arguments{
\item{dat}{matrix of time series (time in rows, variables in columns)}

\item{npc}{number of new dimensions}

\item{met}{selected ITA method}

\item{sca}{whether columns of the original data are scaled (default =TRUE)}

\item{logt}{whether the culumns are log-transformed (default=FALSE)}

\item{metrand}{define the method to create random dataset (either "phase", "rand", or "trend")}

\item{nrep}{number of randomly generated dataset}
}
\value{
percentage variance explained by the successive new dimensions
}
\description{
Extension of ita.rand to generate multiple datasets
}
\examples{
data(baltic)
#matrice of eigen values with multiple randomly generated dataset 
mrand <- ita.nrand(baltic)

# To be compared with eigen values with real observations
mvar <- ita(baltic, npc=2)
plot_nrand(mvar, mrand)
}
