% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita.R
\name{ita.pca}
\alias{ita.pca}
\title{Internal function to run a Principal Component analysis}
\usage{
ita.pca(dat, npc)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of selected principal components}
}
\value{
Result of multivariate analysis with 
\itemize{
\item \code{ts} scores of the time on the principal component
\item \code{co} scores of the variables on the principal component
\item \code{eig} eigen values
\item \code{dat} original dataset
\item \code{npc} number of selected principal component
\item \code{mita} method used for integrated trend analysis 
}
}
\description{
Internal function to run a Principal Component analysis
}
