% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{ita.rand}
\alias{ita.rand}
\title{Run Integrated Trend Analysis on randomly generated dataset}
\usage{
ita.rand(
  dat,
  npc = 2,
  met = "pca",
  sca = TRUE,
  logt = FALSE,
  metrand = "phase"
)
}
\arguments{
\item{dat}{matrix of time series (time in rows, variables in columns)}

\item{npc}{number of new dimensions}

\item{met}{selected ITA method}

\item{sca}{whether columns of the original data are scaled (default =TRUE)}

\item{logt}{whether the culumns are log-transformed (default=FALSE)}

\item{metrand}{define the method to create random dataset (either "phase", "rand", or "trend")}
}
\value{
percentage variance explained by the successive new dimensions
}
\description{
Run Integrated Trend Analysis on randomly generated dataset
}
\examples{
data(baltic)
#Eigen values with random data of similar size than 'baltic'
mrand <- ita.rand(baltic)
}
\seealso{
\code{\link{createRand}}, \code{\link{ita.nrand}}
}
