% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita.R
\name{ita}
\alias{ita}
\title{Run Integrated Trend Analysis}
\usage{
ita(dat, npc, met = "pca", sca = TRUE, logt = FALSE, ...)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of new dimensions}

\item{met}{selected ITA method}

\item{sca}{whether columns of the original data are scaled (default =TRUE)}

\item{logt}{whether the culumns are log-transformed (default=FALSE)}

\item{...}{additional information}
}
\value{
Output of ITA with 
\itemize{
\item \code{ts} scores of the time series on the new dimensions 
\item \code{co} scores of the variables on the new dimensions (if any)
\item \code{eig} eigen values (for PCA associated methods)
\item \code{dat} dataset used for ITA
\item \code{npc} number of new dimensions
\item \code{mita} method used for integrated trend analysis 
}
}
\description{
Run Integrated Trend Analysis
}
\details{
Variable \code{met} are either string or number indicating: 
\itemize{
\item \code{ita.pca} scores of the time series on the new dimensions 
\item \code{ita.pcaV} scores of the variables on the new dimensions (if any)
\item \code{ita.} eigen values (for PCA associated methods)
\item \code{ita.} dataset used for ITA
\item \code{ita.} number of new dimensions
\item \code{ita.} method used for integrated trend analysis 
}
}
\examples{
# Load data
data(baltic)
#  Run ita
mvar <- ita(baltic, npc=2)
# Plot eigen values
plot_eig(mvar)
# Plot scores of years
plot_timeseries(mvar)
# Plot scores of variables
plot_var(mvar)

}
