% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita.R
\name{ita.sfa}
\alias{ita.sfa}
\title{Run Slow Feature Analysis}
\usage{
ita.sfa(dat, npc)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of new dimensions}
}
\value{
Output of ITA with 
\describe{
\item \code{ts} scores of the time series on the new dimensions 
\item \code{co} NULL - no scores of the variables on the new dimensions 
\item \code{eig} NULL 
\item \code{dat} original dataset
\item \code{npc} number of new dimensions
\item \code{mita} method used for integrated trend analysis 
}
}
\description{
Run Slow Feature Analysis
}
\note{
Use function from a package.
}
\examples{

}
\references{
Wiskott L, Sejnowski TJ (2002) Slow Feature Analysis: Unsupervised Learning of Invariances. Neural Comput 14:715–770
Blaschke T, Berkes P, Wiskott L (2006) What Is the Relation Between Slow Feature Analysis and Independent Component Analysis? Neural Comput 18:2495–2508
}
