% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita.R
\name{ita.tsfa}
\alias{ita.tsfa}
\title{Run a Time-series Factor Analysis}
\usage{
ita.tsfa(dat, npc)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of processes}
}
\value{
Output of ITA with 
\describe{
\item \code{ts} scores of the time on the principal component
\item \code{co} scores of the variables on the principal component
\item \code{eig} eigen values
\item \code{dat} original dataset
\item \code{npc} number of selected principal component
\item \code{mita} method used for integrated trend analysis 
}
}
\description{
Run a Time-series Factor Analysis
}
\note{
Use function estTSF.ML from tsfa package.
}
\examples{

}
\references{
Gilbert PD, & Meijer E (2005) Time series factor analysis with an application to measuring money. University of Groningen, Research School SOM Research Report 05F10.
}
