% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{maf}
\alias{maf}
\title{Min/Max Autocorrelation Factors (MAF) in time from  Min/Max Autocorrelation Factors Analysis (MAFA)}
\usage{
maf(formula, data)
}
\arguments{
\item{formula}{A formula with no response variable, referring only to numeric variables.}

\item{data}{A data frame containing the variables in the formula.}
}
\value{
The present routine makes twice use of PCA. Here 'prcomp' is used for PCA
 instead of 'princomp', as it allows less samples (e.g. years) than variables.
 After original variables are turned into normalized PC, MAF are obtained as
 the PC of their increments; their order is then reversed to begin with the
 smallest variograms (i.e. the half variances of increments).
 The output of \code{maf} is  structured similarly than the output from prcomp.
}
\description{
The MAF are a multivariate statistical method, that allows the set of
initial variables to be decomposed into factors. The autocorrelation of
which decreases, or the variogram increases, when going from the first
factors to the last ones. Hence the very first factors extract the part of
the variables which is the most continuous in time.
}
\note{
The number of MAF cannot exceed the number of variables, nor the number of
year increments (number of sampled years - 1). If the number of variables
tends to be larger than the number of sampled years, the MAF no.i (i=1, ...)
tends to have period (nb years - 1)*2/i. In particular there will be
evidence of a high continuity with period (nb years -1)*2 for MAF1,
(nb years -1) for MAF2, etc, whatever the data, which may not be significant
out of this series e.g. for additional years. The number of variables should
then be better reduced well below the number of sampled years.
}
\references{
Solow AR (1994) Detecting Change in the Composition of a Multispecies Community. Biometrics 50:556
Woillez M, Rivoirard J, Petitgas P (2009) Using min/max autocorrelation factors of survey-based indicators to follow the evolution of fish stocks in time. Aquat Living Resour 22:193–200
}
\seealso{
\code{\link{ita.mafa}}
}
\author{
M.Woillez and J. Rivoirard
}
\keyword{Analysis}
\keyword{Autocorrelation}
\keyword{Factors}
\keyword{Min/Max}
