% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multita.R
\name{multita}
\alias{multita}
\title{Run multiple Integrated Trend Analysis}
\usage{
multita(dat, npc, met = c(1, 2), sca = TRUE, logt = FALSE,
  detrend = FALSE)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of new dimensions}

\item{met}{selected ITA methods (can be a vector of multiple methods)}

\item{sca}{whether columns of the original data are scaled (default =TRUE)}

\item{logt}{whether the culumns are log-transformed (default=FALSE)}
}
\value{
A list of output from \code{ita} with 
\describe{
\item \code{ts} scores of the time series on the new dimensions 
\item \code{co} scores of the variables on the new dimensions (if any)
\item \code{eig} eigen values (for PCA associated methods)
\item \code{dat} dataset used for ITA
\item \code{npc} number of new dimensions
\item \code{mita} method used for integrated trend analysis 
}
Additionnally, information about computing time is saved in \code{cptime}
}
\description{
Run multiple Integrated Trend Analysis
}
\details{
Variable \code{met} are either string or number indicating: 
\describe{
\item \code{ita.pca} eeee
\item \code{ita.pcaV} scores of the variables on the new dimensions (if any)
\item \code{ita.} eigen values (for PCA associated methods)
\item \code{ita.} dataset used for ITA
\item \code{ita.} number of new dimensions
\item \code{ita.} method used for integrated trend analysis 
}
}
\note{
The computation can be long, depending on the selected methods.
}
\keyword{methods}
\keyword{multivariate}
