% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{negPC}
\alias{negPC}
\title{Homogenized the sign of PC trends to negative trends}
\usage{
negPC(mvar)
}
\arguments{
\item{mvar}{results of integrative trend analysis (from ita function)}
}
\value{
same object as input, but with scores multiplied by -1 if needed to get a negative trend
}
\description{
Homogenized the sign of PC trends to negative trends
}
