% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{panel.cor.m}
\alias{panel.cor.m}
\title{Function for plotting pairwize correlation}
\usage{
panel.cor.m(x, y, digits = 2)
}
\arguments{
\item{x, y}{x and y}
}
\value{
Pearson correlation
}
\description{
Function for plotting pairwize correlation
}
\note{
Inspired from https://pbil.univ-lyon1.fr/
}
