% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{pc_test}
\alias{pc_test}
\title{Compute p-values for principal component eigen values}
\usage{
pc_test(dat, nrep = 100, metrand = "rand", sca = TRUE,
  logt = FALSE)
}
\arguments{
\item{dat}{matrix with}

\item{nrep}{number of randomly generated dataset}

\item{metrand}{define the method to create random dataset (either "phase", "rand", or "trend")}

\item{sca}{whether columns of the original data are scaled (default =TRUE)}

\item{logt}{whether the culumns are log-transformed (default=FALSE)}
}
\value{
p-values of eigen values
}
\description{
Compute p-values for principal component eigen values
}
\details{
Variable \code{met} can be "phase", "rand" or "trend": 
\itemize{
\item \code{phase} keep similar phase structure in the dataset
\item \code{rand} randomly generated data with similar mean and sd
\item \code{trend} randomly generated data with similar trend
}
}
\examples{
data(baltic)
# p-values of eigen values using randomly generated dataset
pc_test(baltic)
}
\seealso{
\code{\link{createRand}}
}
