% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.compvar}
\alias{plot.compvar}
\title{Visually compare the scores of the variables on reduced dimensions
from different methods of Integrated Trend Analysis}
\usage{
plot.compvar(dat, multimvar, col = brewer.pal(length(multimvar), "Set2"),
  pc = 1, showleg = TRUE, shortname = short(colnames(dat)))
}
\arguments{
\item{dat}{original dataset}

\item{multimvar}{results of multiple ita (from multita function)}

\item{col}{pallette of colors for the different methods of Integrated Trend Analysis}

\item{pc}{which dimension to plot (default pc=1)}

\item{showleg}{whether to plot a legend (default showleg = TRUE)}

\item{shortname}{short name of the variables, use for plotting only}
}
\value{
The pairwize correlation of variable scores compared to the first ITA of \code{multimvar}.
}
\description{
Visually compare the scores of the variables on reduced dimensions
from different methods of Integrated Trend Analysis
}
\author{
Romain Frelat, \email{rom.frelat@gmail.com}
}
