% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eig}
\alias{plot.eig}
\title{Plot eigen values from multivariate analysis}
\usage{
plot.eig(mvar, leg = TRUE, ...)
}
\arguments{
\item{mvar}{results of multivariate analysis (from ita function)}

\item{xax}{aestetic aspect, to plot xaxis on top of the graph}
}
\value{
A barplot of eigen vectors with percentage of variance explained in the successive components.
Also return an object with 
\itemize{
\item \code{barx} the x-positions of the bars in the ploting window
\item \code{perc} percentage of variance explained in the successive PC
}
}
\description{
Plot eigen values from multivariate analysis
}
\author{
Romain Frelat, \email{rom.frelat@gmail.com}
}
\keyword{series}
\keyword{time}
