% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eigrand}
\alias{plot.eigrand}
\title{Compare the eigen values with randomly created matrices.}
\usage{
plot.eigrand(mvar, eigrand, col = brewer.pal(length(R), "Dark2"))
}
\arguments{
\item{mvar}{eigen vectors from ita on original dataset (from ita function)}

\item{eigrand}{eig results from randomly created matrices (from ita.nrand function)}

\item{col}{colors for the significance interval}
}
\value{
Plot of eigen values with 95% confidence interval.
}
\description{
Compare the eigen values with randomly created matrices.
}
\note{
inspired from ade4 r-package
}
\author{
Romain Frelat, \email{rom.frelat@gmail.com}
}
