% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.heatmap}
\alias{plot.heatmap}
\title{Create a heatmap}
\usage{
plot.heatmap(mvar, col, n = 1, shortname = short(colnames(mvar$dat)))
}
\arguments{
\item{mvar}{results of multivariate analysis (from ita function)}

\item{col}{palette of colors}

\item{n}{selected dimension to order the variables (default n=1)}

\item{shortname}{short name of the variables, for plotting}
}
\value{
A heatmap
}
\description{
Create a heatmap
}
\author{
Romain Frelat, \email{rom.frelat@gmail.com}
}
\keyword{heatmap}
