% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.nrand}
\alias{plot.nrand}
\title{Compare the eigen values with randomly created matrices.}
\usage{
plot.nrand(mvar, reig, col = "blue")
}
\arguments{
\item{mvar}{results of ita on original dataset (from ita function)}

\item{reig}{eig results from randomly created matrices (from ita.nrand function)}

\item{col}{color of the 95% confidence interval of \code{reig}}
}
\value{
Plot of eigen values with 95% confidence interval.
}
\description{
Compare the eigen values with randomly created matrices.
}
