% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.randtest}
\alias{plot.randtest}
\title{Compare the eigen values with randomly created matrices.}
\usage{
plot.randtest(obs, distri, nclass = 10, coeff = 1, npc = 1, ...)
}
\arguments{
\item{obs}{eigen vectors from ita on original dataset (from ita function)}

\item{distri}{eig results from randomly created matrices (from ita.nrand function)}

\item{nclass}{number of bars in the histogram}

\item{coeff}{default =1}

\item{npc}{indicating which dimension to compare (default npc = 1)}
}
\value{
Plot comparing the distribution of eigen values from random matrices, with the observed value.
}
\description{
Compare the eigen values with randomly created matrices.
}
\note{
inspired from ade4 r-package
}
\author{
Romain Frelat, \email{rom.frelat@gmail.com}
}
