% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.timeseries}
\alias{plot.timeseries}
\title{Plot time series of scores on the reduced dimensions}
\usage{
plot.timeseries(mvar, xax = 3)
}
\arguments{
\item{mvar}{results of multivariate analysis (from ita function)}

\item{xax}{aestetic aspect, to plot xaxis on top of the graph}
}
\value{
A plot of time series
}
\description{
Plot time series of scores on the reduced dimensions
}
\author{
Romain Frelat, \email{rom.frelat@gmail.com}
}
\keyword{series}
\keyword{time}
