% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.var}
\alias{plot.var}
\title{Plot the scores of variables on reduced dimensions}
\usage{
plot.var(mvar)
}
\arguments{
\item{mvar}{results of multivariate analysis (from ita function)}
}
\value{
A plot of scores of variables on on reduced dimensions
}
\description{
Plot the scores of variables on reduced dimensions
}
\note{
Use s.arrow from ade4 r-package if more than one axis, else a dotchart()
}
\author{
Romain Frelat, \email{rom.frelat@gmail.com}
}
\keyword{series}
\keyword{time}
