% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_compts}
\alias{plot_compts}
\title{Visualy compare the scores of the time series on reduced dimensions
from different methods of Integrated Trend Analysis}
\usage{
plot_compts(dat, multimvar, col = brewer.pal(length(multimvar), "Set2"),
  pc = 1, showleg = TRUE)
}
\arguments{
\item{dat}{original dataset}

\item{multimvar}{results of multiple ita (from multita function)}

\item{col}{pallette of colors for the different methods of Integrated Trend Analysis}

\item{pc}{which dimension to plot (default pc=1)}

\item{showleg}{whether to plot a legend (default showleg = TRUE)}
}
\value{
The pairwize correlation of time series score compared with the first ITA of multimvar
}
\description{
Visualy compare the scores of the time series on reduced dimensions
from different methods of Integrated Trend Analysis
}
\examples{
# Load data
data(baltic)
# Run ita
multi <- multita(baltic, npc=2, met=c("pca", "tsfa"))
# Compare scores of time series
plot_compts(baltic, multi)
}
\keyword{series}
\keyword{time}
