% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_compvar}
\alias{plot_compvar}
\title{Visually compare the scores of the variables on reduced dimensions
from different methods of Integrated Trend Analysis}
\usage{
plot_compvar(
  dat,
  multimvar,
  col = NULL,
  pc = 1,
  showleg = TRUE,
  shortname = short(colnames(dat))
)
}
\arguments{
\item{dat}{original dataset}

\item{multimvar}{results of multiple ita (from multita function)}

\item{col}{pallette of colors for the different methods of Integrated Trend Analysis}

\item{pc}{which dimension to plot (default pc=1)}

\item{showleg}{whether to plot a legend (default showleg = TRUE)}

\item{shortname}{short name of the variables, use for plotting only}
}
\value{
The pairwize correlation of variable scores compared to the first ITA of \code{multimvar}.
}
\description{
Visually compare the scores of the variables on reduced dimensions
from different methods of Integrated Trend Analysis
}
\examples{
#' # Load data
data(baltic)
# Run ita
multi <- multita(baltic, npc=2, met=c("pca", "tsfa"))
plot_compvar(baltic, multi)
}
