% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_eig}
\alias{plot_eig}
\title{Plot eigen values from multivariate analysis}
\usage{
plot_eig(mvar, leg = TRUE, cum = FALSE, ...)
}
\arguments{
\item{mvar}{results of multivariate analysis (from ita function)}

\item{leg}{whether legend is plotted (default = TRUE)}

\item{cum}{whether cumulative variance explained is plotted (default = FALSE)}

\item{...}{other argument for plot}
}
\value{
A barplot of eigen vectors with percentage of variance explained in the successive components.
Also return an object with 
\itemize{
\item \code{barx} the x-positions of the bars in the ploting window
\item \code{perc} percentage of variance explained in the successive PC
}
}
\description{
Plot eigen values from multivariate analysis
}
\examples{
# Load data
data(baltic)
# Run ita
mvar <- ita(baltic, npc=2)
# Plot scores of years
plot_eig(mvar)
}
\keyword{series}
\keyword{time}
