% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Create a heatmap}
\usage{
plot_heatmap(
  mvar,
  col = c("springgreen4", "chartreuse3", "yellow", "darkgoldenrod1", "red"),
  n = 1,
  shortname = short(colnames(mvar$dat))
)
}
\arguments{
\item{mvar}{results of multivariate analysis (from ita function)}

\item{col}{palette of colors (by default, traffic light colors)}

\item{n}{selected dimension to order the variables (default n=1)}

\item{shortname}{short name of the variables, for plotting}
}
\value{
A heatmap
}
\description{
Create a heatmap
}
\examples{
data(baltic)
# Run ita
mvar <- ita(baltic, npc=2)
# Create heatmap
plot_heatmap(mvar)
}
\author{
Romain Frelat, \email{rom.frelat@gmail.com}
}
\keyword{heatmap}
