% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nrand}
\alias{plot_nrand}
\title{Compare the eigen values with randomly created matrices.}
\usage{
plot_nrand(mvar, reig, col = "blue", cum = FALSE)
}
\arguments{
\item{mvar}{results of ita on original dataset (from ita function)}

\item{reig}{eig results from randomly created matrices (from ita.nrand function)}

\item{col}{color of the 95\% confidence interval of \code{reig}}

\item{cum}{whether cumulative variance explained is plotted (default = FALSE)}
}
\value{
Plot of eigen values with 95\% confidence interval.
}
\description{
Compare the eigen values with randomly created matrices.
}
\examples{
# Load data
data(baltic)
# Run ita
mvar <- ita(baltic, npc=2)
# Eigen values with multiple randomly generated data
mrand <- ita.nrand(baltic)
# Graphic comparing eigen values
plot_nrand(mvar, mrand)

# Or with multiple null model
randlist <- list("rand"=ita.nrand(baltic, metrand = "rand"),
                 "phase"=ita.nrand(baltic, metrand = "phase"))
plot_nrand(mvar, randlist, col=c("red", "blue"))
}
