% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time series of scores on the reduced dimensions}
\usage{
plot_timeseries(mvar, xax = 1)
}
\arguments{
\item{mvar}{results of multivariate analysis (from ita function)}

\item{xax}{aestetic aspect, to plot xaxis on top of the graph}
}
\value{
A plot of time series
}
\description{
Plot time series of scores on the reduced dimensions
}
\examples{
# Load data
data(baltic)
#  Run ita
mvar <- ita(baltic, npc=2)
# Plot scores of years
plot_timeseries(mvar)
}
\author{
Romain Frelat, \email{rom.frelat@gmail.com}
}
\keyword{series}
\keyword{time}
