% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{singleRand}
\alias{singleRand}
\title{Create a random time series with similar characteristics than original data}
\usage{
singleRand(ts, metrand = c("rand", "phase", "trend"))
}
\arguments{
\item{ts}{single time series}

\item{metrand}{define the method to create random dataset (phase for keeping phase structure, random data, or random + trend (default = TRUE)}
}
\value{
a randomly generated time series
}
\description{
Create a random time series with similar characteristics than original data
}
\details{
Variable \code{metrand} can be "phase", "rand" or "trend": 
\itemize{
\item \code{phase} keep similar phase structure in the dataset
\item \code{rand} randomly generated data with similar mean and sd
\item \code{trend} randomly generated data with similar trend
}
}
\examples{
data(baltic)
singleRand(baltic[,3], metrand="phase")
}
\seealso{
\code{\link{createRand}}
}
