% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{trend}
\alias{trend}
\title{Compute the linear trend of a time series}
\usage{
trend(ts)
}
\arguments{
\item{ts}{time series}
}
\value{
value of the linear trend
}
\description{
Compute the linear trend of a time series
}
