% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comita.R
\docType{package}
\name{comita}
\alias{comita}
\alias{comita-package}
\title{comita: Comparative tools for Ecosystem Integrative Trend Analysis}
\description{
comita: Comparative tools for Ecosystem Integrative Trend Analysis
}
