% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ita.R
\name{ita.mds}
\alias{ita.mds}
\title{Run Multidimensional scaling}
\usage{
ita.mds(dat, npc)
}
\arguments{
\item{dat}{matrix with}

\item{npc}{number of processes}
}
\value{
Output of ITA with 
\itemize{
\item \code{ts} scores of the time series on the new dimensions 
\item \code{co} NULL - no scores of the variables on the new dimensions 
\item \code{eig} NULL 
\item \code{dat} original dataset
\item \code{npc} number of selected principal component
\item \code{mita} method used for integrated trend analysis 
}
}
\description{
Run Multidimensional scaling
}
\note{
lag set to 1, which correspond to AR1 processes
Function based on dpca from freqdom package
}
\references{
Ku W, Storer RH, Georgakis C (1995) Disturbance detection and isolation by dynamic principal component analysis. Chemom Intell Lab Syst 30:179–196
Ketelaere B De, Hubert M, Schmitt E (2015) Overview of PCA-based statistical process-monitoring methods for time-dependent, high-dimensional data. J Qual Technol 47:318–335
}
