% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{ita.rand}
\alias{ita.rand}
\alias{ita.nrand}
\title{Run Integrated Trend Analysis on randomly generated dataset}
\usage{
ita.rand(dat, npc = 2, met = 1, sca = TRUE, logt = FALSE,
  ar = TRUE, tr = FALSE)

ita.nrand(dat, npc = 2, met = 1, sca = TRUE, logt = FALSE,
  ar = TRUE, tr = FALSE, nrep = 10)
}
\arguments{
\item{dat}{matrix of time series (time in rows, variables in columns)}

\item{npc}{number of new dimensions}

\item{met}{selected ITA method}

\item{sca}{whether columns of the original data are scaled (default =TRUE)}

\item{logt}{whether the culumns are log-transformed (default=FALSE)}

\item{ar}{whether to keep the autocorrelation structure in randomly generated data (default = TRUE)}

\item{tr}{whether to add a trend in randomly generated data similar to the original one (default = FALSE)}

\item{nrep}{number of randomly generated dataset}
}
\value{
percentage variance explained by the successive new dimensions
}
\description{
Run Integrated Trend Analysis on randomly generated dataset
}
\section{Functions}{
\itemize{
\item \code{ita.nrand}: extension to generate multiple datasets
}}

\seealso{
\code{\link{createRand}}, \code{\link{ita.nrand}}
}
